// src/integration/api/notificationApi/path.ts
export enum PATHS {
    NOTIFICATIONS = "/notification",
    INFO_NOTIFICATION = '/info-notification/',
    GET_INFO_NOTIFICATION = '/notifications/info',
    NOTIFICATION_SUBSCRIBE = "/notification/subscribe",
    NOTIFICATION_PUSH = "/notification/push/:id",
    NOTIFICATION = "/notification/:id",
}